Fixed numerous compilation warnings/errors reported with /W4 Fixed an issue with releasing input devices on Win32 which sometimes led to a segfault on shutdown.

+++ doomsday/engine/portable/include/net_main.h
@@ -173,7 +173,7 @@ typedef struct {
 
     // Client-reported time of the last processed ticcmd.
     // Older or as old tics than this are discarded.
-    byte            runTime;
+    int             runTime;
 
     // Bandwidth rating for connection. Determines how much information
     // can be sent to the client. Determined dynamically.
